+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add bg-pink-darker variante and use it on request confirmation form - #969
 - add active sharing link on request confirmation page - #969
 - add Loco Translate integration to ease translations
+- add new Javascript library node-autocomplete - QS-3
+- integrate node-autocomplete - QS-7 QS-4 QS-5 QS-6
 
 ### Changed
 - update all Javascript Dependencies
@@ -67,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - add Docker credentials for Codeship to prevent Pull Rate Limit
+- node-autocomplete library
 
 ## [2.0.0] - 2021-08-26
 ### Added